refactor(core): move ReadOptions from table to config module
ReadOptions is fundamentally a config bag. Move it to config::read_options
where it naturally belongs alongside HudiReadConfig and HudiConfigs.
Re-export from table module for backward compatibility.
